+++ .github/workflows/ci.yml
@@ -34,9 +34,6 @@ jobs:
           pnpm install --frozen-lockfile
           pnpm build
 
-      - name: Install system dependencies
-        run: sudo apt-get update && sudo apt-get install -y libolm-dev
-
       - name: Set up Go
         uses: actions/setup-go@v5
         with:
@@ -89,9 +86,6 @@ jobs:
         run: |
           pnpm install --frozen-lockfile
           pnpm build
-
-      - name: Install system dependencies
-        run: sudo apt-get update && sudo apt-get install -y libolm-dev
 
       - name: Set up Go
         uses: actions/setup-go@v5
